+++ Model/PatternDatabase/Korf/EdgeG2PatternDatabase.h
@@ -15,7 +15,7 @@ namespace busybin
   class EdgeG2PatternDatabase : public EdgePatternDatabase
   {
   public:
-    uint32_t getDatabaseIndex(const RubiksCube& cube) const;
+    size_t getDatabaseIndex(const RubiksCube& cube) const;
   };
 }
